Test Run

+++ cellular.py
@@ -206,12 +206,12 @@ class World:
             fh = self.height
             starty = 0
         else:
-            starty = (self.height - fh) / 2
+            starty = int((self.height - fh) / 2)
         if fw > self.width:
             fw = self.width
             startx = 0
         else:
-            startx = (self.width - fw) / 2
+            startx = int((self.width - fw) / 2)
 
         self.reset()
         for j in range(fh):
